// packages/frontend/src/components/interface/Income-approch-type.tsx
export interface IncomeApprochType {
  appraisal_cost_approach_improvements: any;
  improvements_total_adjusted_ppsf: any;
  improvements_total_adjusted_cost: any;
  effective_age: string;
  year_remodeled: any;
  year_built: any;
  boolean: any;
  income_notes: any;
  total_monthly_income: string | number;
  other_total_monthly_income: string | number;
  total_annual_income: string | number | any;
  other_total_annual_income: string | number | any;
  total_rent_sq_ft: string | number;
  total_sq_ft: string | number | any;
  other_total_sq_ft: string | number | any;
  vacancy: string | number | any;
  vacant_amount: string | number;
  adjusted_gross_amount: string | number | any;
  total_oe_annual_amount: string | number | any;
  total_oe_gross: string | number | any;
  total_oe_per_square_feet: string | number | any;
  expense_notes: string | number;
  total_net_income: string | number;
  monthly_capitalization_rate: string | number;
  annual_capitalization_rate: string | number;
  net_income: string | number;
  sq_ft_capitalization_rate: string | number;
  high_capitalization_rate: string | number;
  cap_rate_notes: string | number;
  indicated_range_monthly: string | number;
  indicated_range_annual: string | number;
  indicated_range_sq_feet: string | number;
  indicated_psf_monthly: string | number;
  indicated_psf_annual: string | number;
  indicated_psf_sq_feet: string | number;
  comparison_basis: string;
  total_unit: string | number | any;
  total_bed: string | number | any;
  total_oe_per_unit: string | number | any;
  total_oe_per_bed: string | number | any;
  incomeSources: Income[];
  otherIncomeSources: OtherIncome[];
  improvements: Income[];
  operatingExpenses: {
    names: any;
    name: string | number;
    annual_amount: string | number | any;
    percentage_of_gross: string | number | any;
    total_per_sq_ft: string | number | any;
    comments: string | number;
    total_per_unit: string | number | any;
    total_per_bed: string | number | any;
  }[];
}

export type Income = {
  space(space: any): unknown;
  id: boolean | undefined;
  link_overview: undefined;
  sf_area: any;
  adjusted_psf: any;
  adjusted_cost?: any;
  depreciation: string | number | any;
  sub_zone: any;
  adjusted_ppsf: any;
  isEditable: any;
  isDisabled: boolean | undefined;
  sq_ft(sq_ft: any): unknown;
  type: string | number;
  monthly_income: string | number | any;
  annual_income: number | string | any;
  rent_sq_ft: string | number | any;
  sf_source: string | number | any;
  unit: string | number | any;
  rent_bed: string | number | any;
  square_feet:string | number | any;
  comments: string;
};

export type OtherIncome = {
  id: boolean | undefined;
  link_overview: undefined;
  sf_area: any;
  adjusted_psf: any;
  adjusted_cost?: any;
  depreciation: string | number | any;
  sub_zone: any;
  adjusted_ppsf: any;
  isEditable: any;
  isDisabled: boolean | undefined;
  sq_ft(sq_ft: any): unknown;
  type: string | number;
  monthly_income: string | number | any;
  annual_income: number | string | any;
  square_feet: string | number | any;
  comments: string;
};