// packages/frontend/src/pages/evaluation/residential/residential-income-approach/residential-incom-approach-enum.ts
export enum ResidentialIncomeApproachEnum {
 INCOME_APPROACH = 'INCOME APPROACH',
 INFORMATION='Enter information about the income and expenses generated by the property.',
 INCOME_SOURCES='incomeSources',
 ANNUAL_INCOME='annual_income',
 MONTHLY_INCOME='monthly_income',
RENT_SQ_FT='rent_sq_ft',
ANNUAL='annual',
BED='Bed',
UNIT='Unit',
MONTHLY='monthly',
RENT='rent',
BOOLEAN='boolean',
UNITS='UNITS',
BEDS='BEDS',
SQ_FT='SQ.FT.',
TOTAL_MONTHLY_INCOME='total_monthly_income',
TOTAL_ANNUAL_INCOME='total_annual_income',
TOTAL_RENT_SQ_FT='total_rent_sq_ft',
VACANT_AMOUNT='vacant_amount',
ADJUSTED_GROSS_AMOUNT='adjusted_gross_amount',
VACANCY='vacancy',
TOTAL_SQ_FT='total_sq_ft',
SELECT_A_SUBTYPE='--Select a Subtype--',
INCOME='INCOME',
TYPE='type',
MONTHLY_INCOME_HEADER='MONTHLY INCOME',
ANNUAL_INCOME_HEADER='ANNUAL INCOME',
IN_ORDER_TO_EDIT_THIS_FIELD='In order to edit this field, you must first input the',
DOLLAR_PER_ACRE='$/Acre',
RENT_AC_YR='RENT/AC/YR',
RENT_SF_YR='RENT/SF/YR',
HAS_OF_BEDS='# OF BEDS',
RENT_BED='rent_bed',
SQUARE_FEET='square_feet',
COMMENTS='COMMENTS',
OTHER_INCOME_SOURCE='Other Income Source',
TOTALS_AND_AVERAGES='TOTALS & AVERAGES',
TOTAL='Total',
TOTAL_ANNUAL_INCOME_HEADER='Total Annual Income',
AVERAGE='Average',
SF='SF',
AC='AC',
INCOME_NOTES_HEADER='INCOME NOTES',
INCOME_NOTES='income_notes',
ADJUSTED_GROSS='ADJUSTED GROSS',
OTHER_INCOME_SOURCES='otherIncomeSources',
OTHER_TOTAL_MONTHLY_INCOME='other_total_monthly_income',
OTHER_TOTAL_ANNUAL_INCOME='other_total_annual_income',
OTHER_TOTAL_SQ_FT='other_total_sq_ft',
OTHER_INCOME='OTHER INCOME',
TOTAL_OE_ANNUAL_AMMOUNT='total_oe_annual_amount',
TOTAL_NET_INCOME='total_net_income',
OPERATING_EXPENSES='operatingExpenses',
PERCENTAGE_OF_GROSS='percentage_of_gross',
TOTAL_PER_SQ_FT='total_per_sq_ft',
TOTAL_PER_BED='total_per_bed',
TOTAL_PER_UNIT='total_per_unit',
ANNUAL_AMOUNT='annual_amount',
TOTAL_OE_GROSS='total_oe_gross',
TOTAL_OE_PER_SQUARE_FEET='total_oe_per_square_feet',
EXPENSE='EXPENSE',
OPERATING_EXPENSES_HEADER='OPERATING EXPENSES',
ANNUAL_AMMOUNT='ANNUAL AMOUNT',
PERCENTAGE_OF_ADJUSTED_GROSS='% OF ADJUSTED GROSS',
DOLLAR_PER_UNIT='$/UNIT',
DOLLAR_PER_BED='$/BED',
DOLLAR_PER_AC='$/AC',
DOLLAR_PER_SF='$/SF',
NAMES='names',
RESIDENTIAL='residential',
NEW_OPERATING_EXPENSE='New Operating Expense',
TOTAL_EXPENSES='TOTAL EXPENSE',
NET_OPERATIING_INCOME='NET OPERATING INCOME',
EXPENSES_NOTES_HEADER='EXPENSES NOTES',
INDICATED_PSF_MONTHLY='indicated_psf_monthly',
INDICATED_PSF_ANNUAL='indicated_psf_annual',
INDICATED_PSF_UNIT='indicated_psf_unit',
MONTHLY_CAPITALIZATION_RATE='monthly_capitalization_rate',
ANNUAL_CAPITALIZATION_RATE='annual_capitalization_rate',
HIGH_CAPITALIZATION_RATE='high_capitalization_rate',
SQ_FT_CAPITALIZATION_RATE='sq_ft_capitalization_rate',
INDICATED_RANGE_MONTHLY='indicated_range_monthly',
INDICATED_RANGE_ANNUAL='indicated_range_annual',
INDICATED_RANGE_SQ_FT='indicated_range_sq_ft',
NET_INCOME='NET INCOME',
CAP_RATES='CAP RATES',
LOW_CAP_RATE='Low CAP Rate',
MARKET_CAP_RATE='Market CAP Rate',
HIGH_CAP_RATE='High CAP Rate',
INDICATED_VALUE_RANGE='INDICATED VALUE RANGE',
INDICATED_VALUE='INDICATED VALUE',
INDICATED_VALUE_DOLLAR_PER_UNIT='INDICATED VALUE $/UNIT',
INDICATED_VALUE_DOLLAR_PER_BED='INDICATED VALUE $/BED',
LAND_ONLY='land_only',
LAND_ONLY_COMMENTS='Land Only',
PROPERTY_TAXES='Property Taxes',
INSURANCE='Insurance',
MAINTENANCE='Maintenance',
MANAGEMENT='Management',
UTILITIES='Utilities',
}
          